requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Session==0.8.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
redis==5.2.1
gunicorn==23.0.0
pytest==8.3.4
pytest-flask==1.3.0
email-validator==2.2.0
